Update frontend API calls
