+++ RecipeFinderApp/app/templates/results.html
@@ -14,6 +14,7 @@
   ] %}
     <h4>Results</h4>
     <!--{{ proteins }}-->
+    {{ API_Results }}
     {% for l in results %}
     <hr>
